+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog]
   - Personal Details Journey
     - Add 'What is your address?'
     - Add 'What gender does your school's payroll system associate with you?'
+    - Add 'Enter bank account details'
 - Automatically update admin claim tasks with DQT API
 
 ## [Release 089] - 2021-03-05